server/pyproject.toml
[project]
name = "pr-graveyard-server"
version = "1.0.0"
description = "PR analysis backend for PR Graveyard"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "requests>=2.31.0",
    "cachetools>=5.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "httpx>=0.26.0"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"